tidies up and introduces single artist performance dataset2
