+++ gpt2-model/dataset.py
@@ -72,7 +72,10 @@ class LyricsDataset():
             # Merge both databases
             csvFile = lyricsCsvFile.merge(artistsCsvFile[['Artist', 'Genres', 'Popularity', 'Link']], left_on='ALink', right_on='Link', how='inner')
             csvFile = self.__preprocess_lyrics_single_artist(csvFile)
-            self.dataset = Dataset.from_pandas(csvFile).select_columns("Lyric").train_test_split(test_size=self.config["val_size"])
+            if self.performance_evaluation_nremovals == None:
+                self.dataset = Dataset.from_pandas(csvFile).select_columns("Lyric").train_test_split(test_size=self.config["val_size"])
+            else:
+                self.dataset = self.__split_train_custom_eval(csvFile, test_size=self.config["val_size"])
             
 
     def load_dataset_multiple_artists(self):
@@ -192,7 +195,7 @@ class LyricsDataset():
             
             data = data.drop(columns=['ALink','SLink','Link','Popularity'])
             #TODO: Remove this
-            #data = data.drop(data.index[2:-1])
+            #data = data.drop(data.index[9:-1])
         return data
     
     def __preprocess_lyrics(self, data):
